[no ci] Added ResourceNotFound when deleting from search

+++ src/main/kotlin/com/leijendary/domain/sample/SampleSearchService.kt
@@ -53,6 +53,12 @@ class SampleSearchService(
     }
 
     fun delete(id: Long) {
+        val exists = sampleSearchRepository.existsById(id)
+
+        if (!exists) {
+            throw ResourceNotFoundException(id, ENTITY, SOURCE)
+        }
+
         sampleSearchRepository.deleteById(id)
     }
